// animes/types.ts
// Generated by https://quicktype.io

export interface Response {
  data: Data;
}

export interface Data {
  Page: Page;
}

export interface Page {
  pageInfo: PageInfo;
  media: Media[];
}

export interface Media {
  id: number;
  coverImage: CoverImage;
  title: Title;
  synonyms: string[];
}

export interface CoverImage {
  extraLarge: string;
}

export interface Title {
  romaji: string;
  english: string;
}

export interface PageInfo {
  total: number;
  perPage: number;
}